Refactoring Conotroller
