--- base-interface/src/main/java/base/interfaces/member/MemberUpdateController.java
@@ -1,28 +1,0 @@
-package base.interfaces.member;
-
-import base.domain.member.MemberModifier;
-import base.support.PasswordEncoderUtils;
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-public class MemberUpdateController {
-
-  private final MemberModifier memberModifier;
-
-  @PatchMapping("/members/{memberId}")
-  public ResponseEntity updatePassword(@PathVariable @NonNull Long memberId,
-      @RequestBody @NonNull String newPassword) {
-
-    String encodePassword = PasswordEncoderUtils.passwordEncode(newPassword);
-    memberModifier.changePassword(memberId, encodePassword);
-
-    return ResponseEntity.ok().build();
-  }
-}
