+++ base-interface/src/main/java/base/support/validator/PhoneNumberValidator.java
@@ -5,10 +5,10 @@ import javax.validation.ConstraintValidatorContext;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-public class PhoneNumberValidator implements ConstraintValidator<PhoneNum, String> {
+public class PhoneNumberValidator implements ConstraintValidator<PhoneNumber, String> {
 
     @Override
-    public void initialize(PhoneNum constraintAnnotation) {
+    public void initialize(PhoneNumber constraintAnnotation) {
 
     }
