+++ base-interface/src/test/groovy/base/interfaces/member/MemberControllerTest.groovy
@@ -1,7 +1,9 @@
 package base.interfaces.member
 
+import base.domain.member.entity.Member
 import base.domain.member.repository.MemberRepository
 import base.interfaces.member.dto.SignUpMember
+import base.support.PasswordEncoderUtils
 import com.fasterxml.jackson.databind.ObjectMapper
 import io.github.benas.randombeans.EnhancedRandomBuilder
 import org.springframework.beans.factory.annotation.Autowired
@@ -12,13 +14,14 @@ import org.springframework.test.web.servlet.MockMvc
 import spock.lang.Shared
 import spock.lang.Specification
 
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.patch
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
 
 @SpringBootTest
 @AutoConfigureMockMvc
-class MemberCreateControllerTest extends Specification {
+class MemberControllerTest extends Specification {
 
     @Autowired
     MockMvc mockMvc
@@ -29,13 +32,37 @@ class MemberCreateControllerTest extends Specification {
     @Autowired
     MemberRepository memberRepository
 
-    @Shared def enhancedRandom
+    @Shared
+    def enhancedRandom;
 
     def setupSpec() {
         enhancedRandom = EnhancedRandomBuilder.aNewEnhancedRandomBuilder()
                 .stringLengthRange(3,5)
                 .collectionSizeRange(1,3)
                 .build();
+    }
+
+    def setup() {
+        memberRepository.deleteAll()
+    }
+
+    def "비밀번호 변경 api 호출시 성공적으로 비밀번호 변경"() {
+        given:
+        def testMember = enhancedRandom.nextObject(Member, "memberId")
+        def newPassword = "newPassword"
+
+        testMember = memberRepository.save(testMember)
+
+        when:
+        mockMvc.perform(patch("/members/{memberId}", testMember.memberId)
+                .content(newPassword))
+                .andDo(print())
+                .andExpect(status().isOk())
+
+        def result = memberRepository.findById(testMember.memberId).get()
+
+        then:
+        assert PasswordEncoderUtils.matches(newPassword, result.memberPassword)
     }
 
     def "입력 형식을 지키지 않고 회원가입을 시도했을 때 400 에러 발생"() {
@@ -53,14 +80,14 @@ class MemberCreateControllerTest extends Specification {
     def "핸드폰 형식을 지키지 않고 회원가입을 시도했을 때 400 에러 발생"() {
         given:
         def signUpMember = SignUpMember.builder()
-                            .memberAddress("멤버주소")
-                            .memberAge(20)
-                            .memberEmail("mollder@naver.com")
-                            .memberName("이인규")
-                            .memberPassword("password")
-                            .memberSex("남")
-                            .memberPhoneNumber("잘못된 핸드폰 번호")
-                            .build()
+                .memberAddress("멤버주소")
+                .memberAge(20)
+                .memberEmail("mollder@naver.com")
+                .memberName("이인규")
+                .memberPassword("password")
+                .memberSex("남")
+                .memberPhoneNumber("잘못된 핸드폰 번호")
+                .build()
 
         expect:
         mockMvc.perform(post("/members")
@@ -69,5 +96,4 @@ class MemberCreateControllerTest extends Specification {
                 .andDo(print())
                 .andExpect(status().isBadRequest())
     }
-
 }
